move sync commands in Admin
